refactor: 오타난 컴포넌트명 수정, 시멘틱한 태그 적용
