+++ src/component/common/Image.jsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from 'react';
+import React, { useEffect } from 'react';
 import styled from 'styled-components';
 
 function Img({ onCheckLazy, ...props }) {
