+++ src/libs/api/userAPI.js
@@ -1,6 +1,6 @@
 import axios from 'axios';
 
 export const getUsersAPI = async () => {
-  const response = await axios.get('http://localhost:3000/data/user.json');
+  const response = await axios.get('/data/user.json');
   return response.data.data;
 };
